選択データ取得クラスを個別ファイルに移動
