+++ sakura/sakura.vcxproj
@@ -531,6 +531,9 @@
     <ClInclude Include="..\sakura_core\view\CCaret.h" />
     <ClInclude Include="..\sakura_core\view\CEditView.h" />
     <ClInclude Include="..\sakura_core\view\CEditView_Paint.h" />
+    <ClInclude Include="..\sakura_core\view\CGetBoxSelectedData.h" />
+    <ClInclude Include="..\sakura_core\view\CGetLinearSelectedData.h" />
+    <ClInclude Include="..\sakura_core\view\CGetSelectedData.h" />
     <ClInclude Include="..\sakura_core\view\colors\CColorStrategy.h" />
     <ClInclude Include="..\sakura_core\view\colors\CColor_Comment.h" />
     <ClInclude Include="..\sakura_core\view\colors\CColor_Found.h" />
@@ -940,6 +943,9 @@
     <ClCompile Include="..\sakura_core\view\CEditView_Paint_Bracket.cpp" />
     <ClCompile Include="..\sakura_core\view\CEditView_Scroll.cpp" />
     <ClCompile Include="..\sakura_core\view\CEditView_Search.cpp" />
+    <ClCompile Include="..\sakura_core\view\CGetBoxSelectedData.cpp" />
+    <ClCompile Include="..\sakura_core\view\CGetLinearSelectedData.cpp" />
+    <ClCompile Include="..\sakura_core\view\CGetSelectedData.cpp" />
     <ClCompile Include="..\sakura_core\view\colors\CColorStrategy.cpp" />
     <ClCompile Include="..\sakura_core\view\colors\CColor_Comment.cpp" />
     <ClCompile Include="..\sakura_core\view\colors\CColor_Found.cpp" />
