+++ sakura_core/view/CEditView.cpp
@@ -43,9 +43,8 @@
 #include "StdAfx.h"
 #include <limits.h>
 #include "view/CEditView.h"
-
-#include <tuple>
-
+#include "view/CGetBoxSelectedData.h"
+#include "view/CGetLinearSelectedData.h"
 #include "view/CViewFont.h"
 #include "view/CRuler.h"
 #include "uiparts/CWaitCursor.h"
@@ -1808,95 +1807,6 @@ bool CEditView::GetSelectedDataSimple( CNativeW &cmemBuf ) const
 	return GetSelectedData( cmemBuf );
 }
 
-/*!
- *  選択範囲のデータを取得するクラス
- */
-class CGetSelectedData
-{
-	using Me = CGetSelectedData;
-
-public:
-	CGetSelectedData() noexcept = default;
-	CGetSelectedData(const Me&) = delete;
-	Me& operator = (const Me&) = delete;
-	virtual ~CGetSelectedData() noexcept = default;
-
-	virtual bool GetData( CNativeW& cmemBuf ) const;
-
-protected:
-	virtual size_t _CountData() const = 0;
-	virtual bool _GetData( CNativeW& cmemBuf ) const = 0;
-};
-
-/*!
- *  通常選択範囲のデータを取得するクラス
- */
-class CGetLinearSelectedData : public CGetSelectedData
-{
-	const CEditView* m_pcEditView;
-	CLayoutPoint ptSelectFrom;
-	CLayoutPoint ptSelectTo;
-	std::wstring_view quoteMark = L"";
-	size_t nLineNumCols = 0;
-	bool bInsertEolAtWrap = false;
-	EEolType newEolType = EEolType::none;
-
-public:
-	explicit CGetLinearSelectedData(
-		const CEditView* m_pcEditView,
-		const CViewSelect& cSelection,
-		std::wstring_view quoteMark,
-		size_t nLineNumCols,
-		bool bInsertEolAtWrap,
-		EEolType newEolType
-	) noexcept;
-
-protected:
-	[[nodiscard]] size_t _CountData() const override;
-	[[nodiscard]] bool _GetData( CNativeW& cmemBuf ) const  override;
-
-private:
-	[[nodiscard]] std::tuple<CLogicXInt, CLogicXInt> LineColumnsToIndexes(const CLayoutInt nLineNum, const CLayout* pcLayout) const {
-		// 行内の桁位置を行頭からのオフセットに変換
-		const auto nIdxFrom = nLineNum == ptSelectFrom.y
-			? m_pcEditView->LineColumnToIndex(pcLayout, ptSelectFrom.x)
-			: CLogicInt(0);
-		const auto nIdxTo = nLineNum == ptSelectTo.y
-			? m_pcEditView->LineColumnToIndex(pcLayout, ptSelectTo.x)
-			: pcLayout->GetLengthWithEOL();
-		return std::make_tuple(nIdxFrom, nIdxTo);
-	}
-};
-
-/*!
- *  矩形選択範囲のデータを取得するクラス
- */
-class CGetBoxSelectedData : public CGetSelectedData
-{
-	const CEditView* m_pcEditView;
-	CLayoutRect rcSel;
-	bool bEnableExtEol;
-
-public:
-	explicit CGetBoxSelectedData(
-		const CEditView* m_pcEditView,
-		const CLayoutRect& _rcSel,
-		const bool _bEnableExtEol
-	) noexcept;
-
-protected:
-	[[nodiscard]] size_t _CountData() const override;
-	[[nodiscard]] bool _GetData( CNativeW& cmemBuf ) const  override;
-
-private:
-	[[nodiscard]] std::tuple<CLogicXInt, CLogicXInt> LineColumnsToIndexes(const CLayoutInt nLineNum, const CLayout* pcLayout) const {
-		// 行内の桁位置を行頭からのオフセットに変換
-		const auto nIdxFrom		= m_pcEditView->LineColumnToIndex( pcLayout, rcSel.left );
-		const auto nIdxTo		= m_pcEditView->LineColumnToIndex( pcLayout, rcSel.right );
-		return std::make_tuple(nIdxFrom, nIdxTo);
-	}
-};
-
 /* 選択範囲のデータを取得
 	正常時はTRUE,範囲未選択の場合はFALSEを返す
 */
@@ -1931,92 +1841,6 @@ bool CEditView::GetSelectedData(
 	}
 }
 
-CGetBoxSelectedData::CGetBoxSelectedData(
-	const CEditView* _pcEditView,
-	const CLayoutRect& _rcSel,
-	const bool _bEnableExtEol
-) noexcept
-	: m_pcEditView(_pcEditView)
-	, rcSel(_rcSel)
-	, bEnableExtEol(_bEnableExtEol)
-{
-}
-
-[[nodiscard]] size_t CGetBoxSelectedData::_CountData() const
-{
-	// コピーに必要なバッファサイズ
-	size_t nBufSize = 0;
-
-	const auto& cLayoutMgr = m_pcEditView->m_pcEditDoc->m_cLayoutMgr;
-
-	// データ計測部
-	for( auto nLineNum = rcSel.top; nLineNum <= rcSel.bottom; ++nLineNum ){
-		const CLayout* pcLayout = nullptr;
-		CLogicInt nLineLen;
-		const auto* pLine = cLayoutMgr.GetLineStr( nLineNum, &nLineLen, &pcLayout );
-		if( !pLine || !pcLayout ){
-			break;
-		}
-
-		// 行データが1文字以上ある場合
-		if (auto [nIdxFrom, nIdxTo] = LineColumnsToIndexes(nLineNum, pcLayout);
-			nIdxFrom < nIdxTo) {
-			// 選択範囲が改行コードで終わっているとき
-			if( WCODE::IsLineDelimiter( pLine[nIdxTo - 1], bEnableExtEol ) ){
-				// 行データの終端は改行コードの手前までにする
-				CEol cEol;
-				cEol.SetTypeByString( &pLine[nIdxFrom], nIdxTo - nIdxFrom );
-				nBufSize += nIdxTo - nIdxFrom - cEol.GetLen();
-			}
-			// 選択範囲が改行コードで終わっていないとき
-			else {
-				nBufSize += nIdxTo - nIdxFrom;
-			}
-		}
-
-		// 矩形選択のコピー時は改行コード固定。
-		nBufSize += 2; // countof(WCODE::CRLF) - 1
-	}
-
-	return nBufSize;
-}
-
-[[nodiscard]] bool CGetBoxSelectedData::_GetData( CNativeW& cmemBuf ) const
-{
-	const auto& cLayoutMgr = m_pcEditView->m_pcEditDoc->m_cLayoutMgr;
-
-	// データ取得部
-	for( auto nLineNum = rcSel.top; nLineNum <= rcSel.bottom; ++nLineNum ){
-		const CLayout* pcLayout = nullptr;
-		CLogicInt nLineLen;
-		const auto* pLine = cLayoutMgr.GetLineStr( nLineNum, &nLineLen, &pcLayout );
-		if( pcLayout == nullptr || pLine == nullptr ){
-			break;
-		}
-
-		// 行データが1文字以上ある場合
-		if (auto [nIdxFrom, nIdxTo] = LineColumnsToIndexes(nLineNum, pcLayout);
-			nIdxFrom < nIdxTo) {
-			// 選択範囲が改行コードで終わっているとき
-			if( WCODE::IsLineDelimiter(pLine[nIdxTo - 1], bEnableExtEol) ){
-				// 行データの終端は改行コードの手前までにする
-				CEol cEol;
-				cEol.SetTypeByString( &pLine[nIdxFrom], nIdxTo - nIdxFrom);
-				cmemBuf.AppendString( &pLine[nIdxFrom], nIdxTo - nIdxFrom - cEol.GetLen() );
-			}
-			// 選択範囲が改行コードで終わっていないとき
-			else{
-				cmemBuf.AppendString( &pLine[nIdxFrom], nIdxTo - nIdxFrom );
-			}
-		}
-
-		// 矩形選択のコピー時は改行コード固定。
-		cmemBuf.AppendString(WCODE::CRLF);
-	}
-
-	return true;
-}
-
 bool CEditView::_GetBoxSelectedData( CNativeW& cmemBuf, const CViewSelect& cSelection, bool bEnableExtEol ) const
 {
 	// 大前提
@@ -2032,151 +1856,6 @@ bool CEditView::_GetBoxSelectedData( CNativeW& cmemBuf, const CViewSelect& cSele
 
 	auto selectedData = CGetBoxSelectedData( this, rcSel, bEnableExtEol );
 	return selectedData.GetData( cmemBuf );
-}
-
-CGetLinearSelectedData::CGetLinearSelectedData(
-	const CEditView* _pcEditView,
-	const CViewSelect& cSelection,
-	std::wstring_view _quoteMark,
-	size_t _nLineNumCols,
-	bool _bInsertEolAtWrap,
-	EEolType _newEolType
-) noexcept
-	: m_pcEditView(_pcEditView)
-	, ptSelectFrom(cSelection.m_sSelect.GetFrom())
-	, ptSelectTo(cSelection.m_sSelect.GetTo())
-	, quoteMark(_quoteMark)
-	, nLineNumCols(_nLineNumCols)
-	, bInsertEolAtWrap(_bInsertEolAtWrap)
-	, newEolType(_newEolType)
-{
-}
-
-[[nodiscard]] size_t CGetLinearSelectedData::_CountData() const
-{
-	// コピーに必要なバッファサイズ
-	size_t nBufSize = 0;
-
-	const auto& cLayoutMgr = m_pcEditView->m_pcEditDoc->m_cLayoutMgr;
-
-	// データ計測部
-	for( auto nLineNum = ptSelectFrom.y; nLineNum <= ptSelectTo.y; ++nLineNum ){
-		const CLayout* pcLayout = nullptr;
-		if( CLogicInt nLineLen(0);
-			!cLayoutMgr.GetLineStr( nLineNum, &nLineLen, &pcLayout ) || !pcLayout ){
-			break;
-		}
-
-		// 行データが1文字以上ある場合
-		if( auto [nIdxFrom, nIdxTo] = LineColumnsToIndexes(nLineNum, pcLayout);
-			nIdxFrom < nIdxTo ){
-			// 引用部分を表す文字列（「> 」など）を付与する
-			if( quoteMark.length() > 0 ){
-				nBufSize += quoteMark.length();
-			}
-
-			// 行番号を付与する
-			if( nLineNumCols > 0 ){
-				nBufSize += nLineNumCols + 2;
-			}
-
-			// 行データが改行コードで終わっているとき
-			if( pcLayout->GetLayoutEol().IsValid() ){
-				nBufSize += nIdxTo - nIdxFrom - pcLayout->GetLayoutEol().GetLen();
-				nBufSize += newEolType == EEolType::none
-					? pcLayout->GetLayoutEol().GetLen()
-					: CEol(newEolType).GetLen();
-			}
-			// 行データが改行コードで終わっていない、かつ、折り返し改行を付けるとき
-			else if (bInsertEolAtWrap){
-				nBufSize += nIdxTo - nIdxFrom;
-				nBufSize += newEolType == EEolType::none
-					? m_pcEditView->m_pcEditDoc->m_cDocEditor.GetNewLineCode().GetLen()
-					: CEol(newEolType).GetLen();
-			}
-			// 行データが改行コードで終わっていないとき
-			else{
-				nBufSize += nIdxTo - nIdxFrom;
-			}
-		}
-	}
-
-	return nBufSize;
-}
-
-[[nodiscard]] bool CGetLinearSelectedData::_GetData( CNativeW& cmemBuf ) const
-{
-	const auto& cLayoutMgr = m_pcEditView->m_pcEditDoc->m_cLayoutMgr;
-
-	// 行番号整形バッファ(L" 1234:"を出力できるよう桁数+2桁分確保する)
-	std::wstring lineNumBuf(nLineNumCols + 2, wchar_t());
-
-	// データ取得部
-	for( auto nLineNum = ptSelectFrom.y; nLineNum <= ptSelectTo.y; ++nLineNum ){
-		const CLayout* pcLayout = nullptr;
-		CLogicInt nLineLen;
-		const auto *pLine = cLayoutMgr.GetLineStr( nLineNum, &nLineLen, &pcLayout );
-		if( pLine == nullptr || pcLayout == nullptr){
-			break;
-		}
-
-		// 行データが1文字以上ある場合
-		if( auto [nIdxFrom, nIdxTo] = LineColumnsToIndexes(nLineNum, pcLayout);
-			nIdxFrom < nIdxTo ){
-			// 引用部分を表す文字列（「> 」など）を付与する
-			if( quoteMark.length() > 0 ){
-				cmemBuf.AppendString( quoteMark.data() );
-			}
-
-			// 行番号を付与する
-			if( nLineNumCols > 0 ){
-				// 行番号は L" 1234:" 形式で出力する
-				::swprintf_s(lineNumBuf.data(), lineNumBuf.capacity(), L"% *d:", static_cast<uint32_t>(nLineNumCols), (int)(Int)(nLineNum + 1));
-				cmemBuf.AppendString(lineNumBuf.data());
-			}
-
-			// 行データが改行コードで終わっているとき
-			if( pcLayout->GetLayoutEol().IsValid() ){
-				// 行データの終端は改行コードの手前までにする
-				cmemBuf.AppendString( &pLine[nIdxFrom], nIdxTo - nIdxFrom - pcLayout->GetLayoutEol().GetLen() );
-				// 変換指定に従い、改行コードを付与する
-				cmemBuf.AppendString(newEolType == EEolType::none
-					? pcLayout->GetLayoutEol().GetValue2()						//	コード保存
-					: CEol(newEolType).GetValue2());							//	新規改行コード
-			}
-			// 行データが改行コードで終わっていない、かつ、折り返し改行を付けるとき
-			else if (bInsertEolAtWrap){
-				// 行データの終端は改行コードの手前までにする
-				cmemBuf.AppendString(&pLine[nIdxFrom], nIdxTo - nIdxFrom);
-				// ドキュメントの改行コードまたは指定された改行コードを付与する
-				cmemBuf.AppendString(newEolType == EEolType::none
-					? m_pcEditView->m_pcEditDoc->m_cDocEditor.GetNewLineCode().GetValue2()	//	コード保存
-					: CEol(newEolType).GetValue2());							//	新規改行コード
-			}
-			// 行データが改行コードで終わっていないとき
-			else{
-				cmemBuf.AppendString(&pLine[nIdxFrom], nIdxTo - nIdxFrom);
-			}
-		}
-	}
-
-	return true;
-}
-
-[[nodiscard]] bool CGetSelectedData::GetData( CNativeW& cmemBuf ) const
-{
-	// コピーに必要なバッファサイズを計測
-	const size_t nBufSize = _CountData();
-
-	// メモリ確保
-	cmemBuf.Clear();
-	cmemBuf.AllocStringBuffer(nBufSize);
-	if( cmemBuf.capacity() < nBufSize ){
-		return false;
-	}
-
-	// データ取得
-	return _GetData( cmemBuf );
 }
 
 bool CEditView::_GetLinearSelectedData( CNativeW& cmemBuf, const CViewSelect& cSelection, std::wstring_view quoteMark, bool bWithLineNumber, bool bInsertEolAtWrap, EEolType newEolType ) const
